Move link definitions into their own file
